+++ wradlib/io/radolan.py
@@ -52,6 +52,45 @@ dwdpattern = re.compile("raa..-(..)[_-]([0-9]{5})-([0-9]*)-(.*?)---bin")
 # RW_20221015-0050.asc
 dwdascii = re.compile("(..)_([0-9]*)-([0-9]*).asc")
 
+dwdradarmapping = {
+    10103: "asb",
+    10132: "boo",
+    10488: "drs",
+    10780: "eis",
+    10410: "ess",
+    10908: "fbg",
+    10440: "fld",
+    10339: "hnr",
+    10873: "isn",
+    10950: "mem",
+    10557: "neu",
+    10605: "nhb",
+    10629: "oft",
+    10392: "pro",
+    10169: "ros",
+    10832: "tur",
+    10356: "umd",
+}
+dwdlocs = {
+    "asb": {"height": 36.22, "lat": 53.564129, "lon": 6.748317},
+    "boo": {"height": 124.56, "lat": 54.004381, "lon": 10.046899},
+    "drs": {"height": 263.36, "lat": 51.124639, "lon": 13.768639},
+    "eis": {"height": 799.08, "lat": 49.540667, "lon": 12.402788},
+    "ess": {"height": 185.11, "lat": 51.405649, "lon": 6.967111},
+    "fbg": {"height": 1516.1, "lat": 47.873611, "lon": 8.003611},
+    "fld": {"height": 627.88, "lat": 51.311197, "lon": 8.801998},
+    "hnr": {"height": 97.78, "lat": 52.460083, "lon": 9.694533},
+    "isn": {"height": 677.77, "lat": 48.174705, "lon": 12.101779},
+    "mem": {"height": 724.4, "lat": 48.042145, "lon": 10.219222},
+    "neu": {"height": 879.65, "lat": 50.500114, "lon": 11.135034},
+    "nhb": {"height": 585.85, "lat": 50.109656, "lon": 6.548328},
+    "oft": {"height": 245.8, "lat": 49.984745, "lon": 8.712933},
+    "pro": {"height": 193.92, "lat": 52.648667, "lon": 13.858212},
+    "ros": {"height": 37.05, "lat": 54.17566, "lon": 12.058076},
+    "tur": {"height": 767.62, "lat": 48.585379, "lon": 9.782675},
+    "umd": {"height": 185.2, "lat": 52.160096, "lon": 11.176091},
+}
+
 
 def _get_timestamp_from_filename(filename, *, pattern=dwdpattern):
     """Helper function doing the actual work of get_dx_timestamp"""
@@ -373,6 +412,22 @@ def get_radolan_header_token():
     return head
 
 
+def get_radolan_site_token():
+    head = get_radolan_header_token()
+    head.update(
+        {
+            "CO": None,
+            "CD": None,
+            "MH": None,
+            "HI": None,
+            "CI": None,
+            "CL": None,
+            "FL": None,
+        }
+    )
+    return head
+
+
 def get_radolan_header_token_pos(header, *, mode="composite"):
     """Get Token and positions from DWD radolan header
 
@@ -381,21 +436,22 @@ def get_radolan_header_token_pos(header, *, mode="composite"):
     header : str
         (ASCII header)
     mode : str, optional
-        'composite' or 'dx', defaults to 'composite'.
+        'composite', 'station' or 'dx', defaults to 'composite'.
 
     Returns
     -------
     head : dict
         with found header tokens and positions
     """
-
     if mode == "composite":
         head_dict = get_radolan_header_token()
+    elif mode == "station":
+        head_dict = get_radolan_site_token()
     elif mode == "dx":
         head_dict = get_dx_header_token()
     else:
         raise ValueError(
-            f"Unknown mode {mode}, use either `composite` or `dx` depending on data source"
+            f"Unknown mode {mode}, use either `composite`, `stations or `dx` depending on data source"
         )
 
     for token in head_dict.keys():
@@ -450,14 +506,23 @@ def parse_dwd_composite_header(header):
     # radar location ID (always 10000 for composites)
     out["radarid"] = header[8:13]
 
+    # if station product
+    if out["radarid"] == 10000:
+        kwargs = dict(mode="composite")
+    else:
+        kwargs = dict(mode="station")
+        # assume site product with 200 rows/cols
+        out["nrow"] = 200
+        out["ncol"] = 200
+
     # get dict of header token with positions
-    head = get_radolan_header_token_pos(header)
+    head = get_radolan_header_token_pos(header, **kwargs)
     # iterate over token and fill output dict accordingly
     for k, v in head.items():
         if v:
             if k == "BY":
                 out["datasize"] = int(header[v[0] : v[1]]) - len(header) - 1
-            if k == "VS":
+            elif k == "VS":
                 vs = int(header[v[0] : v[1]])
                 out["formatversion"] = vs
                 out["maxrange"] = {
@@ -466,21 +531,21 @@ def parse_dwd_composite_header(header):
                     2: "128 km",
                     3: "150 km",
                 }.get(vs, "100 km")
-            if k == "SW":
+            elif k == "SW":
                 out["radolanversion"] = header[v[0] : v[1]].strip()
-            if k == "PR":
+            elif k == "PR":
                 out["precision"] = float("1" + header[v[0] : v[1]].strip())
-            if k == "INT":
+            elif k == "INT":
                 out["intervalseconds"] = int(header[v[0] : v[1]]) * 60
-            if k == "U":
+            elif k == "U":
                 out["intervalunit"] = int(header[v[0] : v[1]])
                 if out["intervalunit"] == 1:
                     out["intervalseconds"] *= 1440
-            if k == "GP":
+            elif k == "GP":
                 dimstrings = header[v[0] : v[1]].strip().split("x")
                 out["nrow"] = int(dimstrings[0])
                 out["ncol"] = int(dimstrings[1])
-            if k == "BG":
+            elif k == "BG":
                 dimstrings = header[v[0] : v[1]]
                 dimstrings = (
                     dimstrings[: int(len(dimstrings) / 2)],
@@ -488,32 +553,59 @@ def parse_dwd_composite_header(header):
                 )
                 out["nrow"] = int(dimstrings[0])
                 out["ncol"] = int(dimstrings[1])
-            if k == "LV":
+            elif k == "LV":
                 lv = header[v[0] : v[1]].split()
                 out["nlevel"] = int(lv[0])
                 out["level"] = np.array(lv[1:]).astype("float")
-            if k == "MS":
-                locationstring = header[v[0] :].strip().split("<")[1].split(">")[0]
-                out["radarlocations"] = locationstring.split(",")
-            if k == "ST":
+            elif k == "MS":
+                try:
+                    locationstring = header[v[0] :].strip().split("<")[1].split(">")[0]
+                    out["radarlocations"] = locationstring.split(",")
+                except IndexError:
+                    try:
+                        cnt = int(header[v[0] : v[0] + 3])
+                        out["message"] = header[v[0] + 3 : v[0] + 3 + cnt]
+                    except ValueError:
+                        pass
+            elif k == "ST":
                 locationstring = header[v[0] :].strip().split("<")[1].split(">")[0]
                 out["radardays"] = locationstring.split(",")
-            if k == "CS":
-                out["indicator"] = {
-                    0: "near ground level",
-                    1: "maximum",
-                    2: "tops",
-                }.get(int(header[v[0] : v[1]]))
-            if k == "MX":
+            elif k == "CS":
+                if out["producttype"] == "PZ":
+                    out["statisticfilter"] = header[v[0] : v[1]].strip()
+                else:
+                    out["indicator"] = {
+                        0: "near ground level",
+                        1: "maximum",
+                        2: "tops",
+                    }.get(int(header[v[0] : v[1]]))
+            elif k == "MX":
                 out["imagecount"] = int(header[v[0] : v[1]])
-            if k == "VV":
+            elif k == "VV":
                 out["predictiontime"] = int(header[v[0] : v[1]])
-            if k == "MF":
+            elif k == "MF":
                 out["moduleflag"] = int(header[v[0] : v[1]])
-            if k == "QN":
+            elif k == "QN":
                 out["quantification"] = int(header[v[0] : v[1]])
-            if k == "VR":
+            elif k == "VR":
                 out["reanalysisversion"] = header[v[0] : v[1]].strip()
+            elif k == "CO":
+                out["cluttermap"] = header[v[0] : v[1]].strip()
+            elif k == "CD":
+                out["dopplerfilter"] = header[v[0] : v[1]].strip()
+            elif k == "MH":
+                out["maxheight"] = int(header[v[0] : v[1]])
+            elif k == "HI":
+                out["hailwarning"] = header[v[0] : v[1]].strip()
+            elif k == "CI":
+                ci = header[v[0] : v[1]]
+                lci = len(ci) // 2
+                out["severeconvection"] = [float(ci[:lci]), float(ci[lci:])]
+            elif k == "CL":
+                cl = header[v[0] : v[1]]
+                out["severeconvectionheights"] = [int(cl[0:2]), int(cl[2:4])]
+            elif k == "FL":
+                out["freezing_level"] = header[v[0] : v[1]].strip()
     return out
 
 
@@ -629,8 +721,12 @@ def decode_radolan_runlength_array(binarr, attrs):
         dline = decode_radolan_runlength_line(line, attrs)
         arr = np.vstack((arr, dline))
         line = read_radolan_runlength_line(buf)
+
+    # reshape early for PZ station product
+    if mh := attrs.get("maxheight", False):
+        arr = arr.reshape(mh, attrs["nrow"], attrs["ncol"])
     # return upside down because first line read is top line
-    return np.flipud(arr)
+    return np.flip(arr, axis=-2)
 
 
 def read_radolan_binary_array(fid, size, *, raise_on_error=True):
@@ -813,7 +909,7 @@ def read_radolan_composite(f, *, missing=-9999, loaddata=True, fillmissing=False
         arr = arr * attrs["precision"]
     # set nodata value
     if "nodatamask" in attrs:
-        arr.flat[attrs["nodatamask"]] = NODATA
+        arr.flat[attrs["nodatamask"].flatten()] = NODATA
     return arr, attrs
 
 
@@ -833,10 +929,9 @@ def _get_radolan_product_attributes(attrs):
     product = attrs["producttype"]
     pattrs = {}
 
-    if product not in ["PG", "PC", "ascii"]:
+    if product not in ["PG", "PC", "PZ", "ascii"]:
         interval = attrs["intervalseconds"]
         precision = attrs["precision"]
-
     if product in ["RX", "EX", "WX"]:
         pattrs.update(radolan["dBZ"])
     elif product in ["WN"]:
@@ -874,7 +969,7 @@ def _get_radolan_product_attributes(attrs):
     ]:
         pattrs.update(radolan["RA"])
         pattrs.update({"scale_factor": np.float32(precision)})
-    elif product in ["PG", "PC"]:
+    elif product in ["PG", "PC", "PZ"]:
         pattrs.update(radolan["PG"])
     elif product in ["%M", "%J", "%Y"]:
         pattrs.update(radolan["%"])
@@ -1027,7 +1122,10 @@ class _radolan_file:
     @property
     def shape(self):
         if self._shape is None:
-            self._shape = (self.attrs["nrow"], self.attrs["ncol"])
+            shape = (self.attrs["nrow"], self.attrs["ncol"])
+            if mh := self.attrs.get("maxheight", False):
+                shape = (mh,) + shape
+            self._shape = shape
         return self._shape
 
     @property
@@ -1065,9 +1163,9 @@ class _radolan_file:
         data = self._data[self.product]
         if self.attrs["producttype"] == "ascii":
             self.attrs["nodatamask"] = np.where(data == self.attrs["nodataflag"])[0]
-        elif self.product in ["PG", "PC"]:
+        elif self.product in ["PG", "PC", "PZ"]:
             self.attrs["nodataflag"] = 255
-            self.attrs["nodatamask"] = np.where(data == 255)[0]
+            self.attrs["nodatamask"] = data == 255
         elif self.product in ["RX", "EX", "WX"]:
             self.attrs["nodatamask"] = np.where(data == 250)[0]
             self.attrs["cluttermask"] = np.where(data == 249)[0]
@@ -1099,18 +1197,24 @@ class _radolan_file:
 
         # handle truncated data
         binarr_kwargs = {}
-        if self._fill and self.product not in ["PG", "PC"]:
+        if self._fill and self.product not in ["PG", "PC", "PZ"]:
             binarr_kwargs.update({"raise_on_error": False})
         # read data
         size = self.attrs["datasize"]
         indat = read_radolan_binary_array(self.fp, size, **binarr_kwargs)
 
-        if self._fill and len(indat) < size and self.product not in ["PG", "PC"]:
+        if self._fill and len(indat) < size and self.product not in ["PG", "PC", "PZ"]:
             indat = _fix_radolan_truncated_buffer(indat, size, self.dtype)
 
-        if self.product in ["PC", "PG"]:
+        if self.product in ["PC", "PG", "PZ"]:
             self._data[self.product] = decode_radolan_runlength_array(
-                indat, {"ncol": self.attrs["ncol"], "nodataflag": 255}
+                indat,
+                {
+                    "ncol": self.attrs["ncol"],
+                    "nrow": self.attrs["nrow"],
+                    "maxheight": self.attrs.get("maxheight", False),
+                    "nodataflag": 255,
+                },
             )
         else:
             self._data[self.product] = np.frombuffer(indat, dtype=self.dtype)
@@ -1123,16 +1227,19 @@ class _radolan_file:
                 self._data[self.product] = self._data[self.product].astype(self.dtype)
         self._process_data()
         for _k, v in self._data.items():
-            v.shape = (self.attrs["nrow"], self.attrs["ncol"])
+            v.shape = self.shape
+            self._data[_k] = v
 
     def _read(self):
         attrs = self.attrs.copy()
         pattrs = _get_radolan_product_attributes(attrs)
 
+        if self.attrs.get("maxheight", False):
+            self.dimensions["z"] = self.attrs["maxheight"]
         self.dimensions["y"] = self.attrs["nrow"]
         self.dimensions["x"] = self.attrs["ncol"]
-
-        pattrs.update({"long_name": self.product, "coordinates": "time y x"})
+        dims = " ".join(list(self.dimensions.keys()))
+        pattrs.update({"long_name": self.product, "coordinates": f"time {dims}"})
 
         data_var = WradlibVariable(self.dimensions, data=self, attrs=pattrs)
 
@@ -1156,17 +1263,25 @@ class _radolan_file:
 
         # if GDAL is not installed, fall back to trigonometric calculation
         if util.has_import(osr):
-            print("CRS")
             if attrs.get("formatversion", 3) >= 5:
                 crs = projection.create_osr("dwd-radolan-wgs84")
             else:
                 crs = projection.create_osr("dwd-radolan-sphere")
         else:
-            print("TRIG")
             crs = "trig"
 
+        coords_kwargs = {}
+        if self.attrs.get("maxheight", False):
+            radar = dwdradarmapping[int(self.attrs["radarid"])]
+            site = dwdlocs[radar]
+            coords_kwargs.update(dict(lon=site["lon"], lat=site["lat"]))
+
         xlocs, ylocs = rect.get_radolan_coordinates(
-            self.dimensions["y"], self.dimensions["x"], crs=crs, mode="center"
+            self.dimensions["y"],
+            self.dimensions["x"],
+            crs=crs,
+            mode="center",
+            **coords_kwargs,
         )
         xattrs = {
             "units": "m",
@@ -1182,12 +1297,31 @@ class _radolan_file:
         }
         y_var = WradlibVariable("y", ylocs, yattrs)
 
+        if self.attrs.get("maxheight", False):
+            zattrs = {
+                "units": "m",
+                "long_name": "z coordinate of projection",
+                "standard_name": "projection_z_coordinate",
+            }
+            zlocs = np.arange(
+                1000, self.attrs["maxheight"] * 1000 + 1, 1000, dtype=float
+            )
+            z_var = WradlibVariable("z", zlocs, zattrs)
+
         self._variables = {
             self.product: data_var,
             "time": time_var,
-            "y": y_var,
-            "x": x_var,
         }
+
+        if self.attrs.get("maxheight", False):
+            self._variables.update({"z": z_var})
+
+        self._variables.update(
+            {
+                "y": y_var,
+                "x": x_var,
+            }
+        )
 
         if pred_time is not None:
             self._variables.update({"prediction_time": pred_time_var})
@@ -1216,7 +1350,7 @@ class _radolan_file:
         self.attributes.update(attrs)
 
     def _get_ancillary(self, requested=True):
-        if self.product in ["PG", "PC"] or self.attrs["producttype"] == "ascii":
+        if self.product in ["PG", "PC", "PZ"] or self.attrs["producttype"] == "ascii":
             anc = ("nodatamask",)
         elif self.product in ["RX", "EX", "WX"]:
             anc = ("nodatamask", "cluttermask")